Change Place.id to be a long (at the X10 level).  At the native layers, the Place id continues to be an int.
Changes perhaps should ripple futher into Team and Runtime, but will 
leave that to Ben and Olivier to do in later commits as appropriate.



git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@25529 a326200e-df2c-42e0-81fa-597e909af41d

+++ apgas.java/src/com/ibm/apgas/Pool.java
@@ -83,7 +83,7 @@ public class Pool extends x10.runtime.impl.java.Runtime {
      * How many places are there in the current execution?
      */
     public static int numPlaces() { 
-        return x10.lang.Place.numPlaces$O();
+        return (int)x10.lang.Place.numPlaces$O();
     }
     
     /**
